requirements.txt
moviepy==1.0.3
Pillow==10.0.1
tkinter-tooltip==2.0.0
customtkinter==5.2.0
opencv-python==4.8.1.78
numpy==1.24.3
imageio==2.31.5
imageio-ffmpeg==0.4.9
pydub==0.25.1 